Adding JRPG Combat system
made the Scientist and Lab Rat modular systems in which then were implimented to a modular combat system allowing for differing movesets, healths, and abilities all with a base in polymorphism. Restructured Maze as well as all other aspects of the project. This was a complete overhaul of the project.